requirements.txt
better_abc==0.0.3
bitsandbytes==0.43.3
datasets==2.20.0
einops==0.8.1
fancy_einsum==0.0.3
huggingface_hub==0.29.2
jaxtyping==0.3.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
Pillow==11.3.0
pytest==8.4.1
rich==14.0.0
torch==2.4.0
tqdm==4.66.4
transformers==4.47.0
typeguard==2.13.3
typing_extensions==4.14.1
wandb==0.21.0
